+++ xmake/modules/lib/detect/check_cxsnippets.lua
@@ -119,7 +119,8 @@ end
 -- @param opt       the argument options
 --                  .e.g 
 --                  { verbose = false, target = [target|option], sourcekind = "[cc|cxx]"
---                  , types = {"wchar_t", "char*"}, includes = "stdio.h", funcs = {"sigsetjmp", "sigsetjmp((void*)0, 0)"}}
+--                  , types = {"wchar_t", "char*"}, includes = "stdio.h", funcs = {"sigsetjmp", "sigsetjmp((void*)0, 0)"}
+--                  , config = {defines = "xx", cxflags = ""}}
 --
 -- funcs:
 --      sigsetjmp